TrackRecordingService: refactored and unified tests.
